# fe/api-docs-admin.yaml
openapi: 3.0.1
info:
  title: Shop Admin API
  description: API documentation for Shop Admin
  version: v1
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/admin/payments/{id}/refund:
    put:
      tags:
      - payment-admin-proxy-controller
      operationId: refund
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/admin/payments/{id}/mark-paid:
    put:
      tags:
      - payment-admin-proxy-controller
      operationId: markPaid
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: transactionId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/admin/payments/{id}/cancel:
    put:
      tags:
      - payment-admin-proxy-controller
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/admin/orders/{orderId}/status:
    put:
      tags:
      - order-admin-controller
      operationId: updateOrderStatus
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/admin/orders/{orderId}/payment-status:
    put:
      tags:
      - order-admin-controller
      operationId: updatePaymentStatus
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
  /api/admin/orders:
    post:
      tags:
      - order-admin-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/admin/inventory:
    get:
      tags:
      - inventory-admin-controller
      operationId: getAll
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageInventoryResponse"
    post:
      tags:
      - inventory-admin-controller
      operationId: createOrUpdate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
  /api/admin/inventory/reserve:
    post:
      tags:
      - inventory-admin-controller
      operationId: reserve
      parameters:
      - name: variantId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /api/admin/inventory/release:
    post:
      tags:
      - inventory-admin-controller
      operationId: release
      parameters:
      - name: variantId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /api/admin/inventory/import:
    post:
      tags:
      - inventory-admin-controller
      operationId: importStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
  /api/admin/inventory/export:
    post:
      tags:
      - inventory-admin-controller
      operationId: exportStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
  /api/admin/inventory/deduct:
    post:
      tags:
      - inventory-admin-controller
      operationId: deduct
      parameters:
      - name: variantId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /api/admin/inventory/adjust/{variantId}:
    post:
      tags:
      - inventory-admin-controller
      operationId: adjust
      parameters:
      - name: variantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
...
  /api/admin/payments:
    get:
      tags:
      - payment-admin-proxy-controller
      operationId: getAllPayments
      parameters:
      - name: filter
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/PaymentSearchRequest"
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagePaymentResponse"
  /api/admin/payments/{id}:
    get:
      tags:
      - payment-admin-proxy-controller
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/admin/payments/summary:
    get:
      tags:
      - payment-admin-proxy-controller
      operationId: getSummary_1
      parameters:
      - name: fromDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/admin/orders/{orderId}:
    get:
      tags:
      - order-admin-controller
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/admin/inventory/{variantId}:
    get:
      tags:
      - inventory-admin-controller
      operationId: getByVariant
      parameters:
      - name: variantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryResponse"
  /api/admin/inventory/transactions/{variantId}:
    get:
      tags:
      - inventory-admin-controller
      operationId: getTransactions
      parameters:
      - name: variantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransactionDTO"
components:
  schemas:
...
    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        guestId:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - FAILED
          - REFUNDED
          - CANCELLED
        amount:
          type: number
        method:
          type: string
          enum:
          - COD
          - VNPAY
        transactionId:
          type: string
        paidAt:
          type: string
          format: date-time
        paymentUrl:
          type: string
    StatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        variantId:
          type: integer
          format: int64
        size:
          type: string
        color:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
          format: int32
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        guestId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        shippingAddress:
          type: string
        billingAddress:
          type: string
        status:
          type: string
          enum:
          - NEW
          - PENDING_PAYMENT
          - CONFIRMED
          - PROCESSING
          - PACKING
          - SHIPPING
          - DELIVERED
          - COMPLETED
          - CANCELLED
        paymentStatus:
          type: string
          enum:
          - PENDING
          - PAID
          - FAILED
          - REFUNDED
          - CANCELLED
        totalPrice:
          type: number
        shippingFee:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        paymentMethod:
          type: string
        shippingMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        paymentDate:
          type: string
          format: date-time
        estimatedDeliveryDate:
          type: string
          format: date-time
        actualDeliveryDate:
          type: string
          format: date-time
    PaymentUpdateRequest:
      type: object
      properties:
        paymentStatus:
          type: string
...
    CreateOrderRequest:
      required:
      - items
      type: object
      properties:
        userId:
          type: integer
          format: int64
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        shippingAddress:
          type: string
        billingAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemRequest"
    OrderItemRequest:
      required:
      - productId
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          minimum: 1
          type: integer
          format: int32
    InventoryRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        variantId:
          type: integer
          format: int64
        stock:
          type: integer
          format: int32
        importPrice:
          type: number
        sellingPrice:
          type: number
    InventoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        variantId:
          type: integer
          format: int64
        stock:
          type: integer
          format: int32
        reservedQuantity:
          type: integer
          format: int32
        availableQuantity:
          type: integer
          format: int32
        importPrice:
          type: number
        sellingPrice:
          type: number
        color:
          type: string
        size:
          type: string
        productName:
          type: string
        thumbnail:
          type: string
    ImportRequest:
      type: object
      properties:
        variantId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        importPrice:
          type: number
        note:
          type: string
    ExportRequest:
      type: object
      properties:
        variantId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        note:
          type: string
    AdjustRequest:
      type: object
      properties:
        newStock:
          type: integer
          format: int32
        reason:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    PaymentSearchRequest:
      type: object
      properties:
        status:
          type: string
        method:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        minAmount:
          type: number
        maxAmount:
          type: number
        transactionId:
          type: string
        userId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
    PagePaymentResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/PaymentResponse"
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageInventoryResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/InventoryResponse"
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    InventoryTransactionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        variantId:
          type: integer
          format: int64
        type:
          type: string
        quantity:
          type: integer
          format: int32
        note:
          type: string
...