# fe/src/app/api/admin/model/orderResponse.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderItemResponse } from './orderItemResponse';


export interface OrderResponse { 
    id?: number;
    userId?: number;
    guestId?: string;
    guestName?: string;
    guestEmail?: string;
    guestPhone?: string;
    shippingAddress?: string;
    billingAddress?: string;
    status?: OrderResponseStatusEnum;
    paymentStatus?: OrderResponsePaymentStatusEnum;
    totalPrice?: number;
    shippingFee?: number;
    discountAmount?: number;
    finalAmount?: number;
    paymentMethod?: string;
    shippingMethod?: string;
    items?: Array<OrderItemResponse>;
    createdAt?: string;
    updatedAt?: string;
    paymentDate?: string;
    estimatedDeliveryDate?: string;
    actualDeliveryDate?: string;
}
export enum OrderResponseStatusEnum {
    NEW = 'NEW',
    PENDING_PAYMENT = 'PENDING_PAYMENT',
    CONFIRMED = 'CONFIRMED',
    PROCESSING = 'PROCESSING',
    PACKING = 'PACKING',
    SHIPPING = 'SHIPPING',
    DELIVERED = 'DELIVERED',
    COMPLETED = 'COMPLETED',
    CANCELLED = 'CANCELLED'
};
export enum OrderResponsePaymentStatusEnum {
    PENDING = 'PENDING',
    PAID = 'PAID',
    FAILED = 'FAILED',
    REFUNDED = 'REFUNDED',
    CANCELLED = 'CANCELLED'
};