# fe/src/app/api/admin/model/pageableObject.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SortObject } from './sortObject';


export interface PageableObject { 
    offset?: number;
    sort?: Array<SortObject>;
    paged?: boolean;
    pageSize?: number;
    pageNumber?: number;
    unpaged?: boolean;
}