# fe/src/app/api/admin/model/paymentSearchRequest.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentSearchRequest { 
    status?: string;
    method?: string;
    fromDate?: string;
    toDate?: string;
    minAmount?: number;
    maxAmount?: number;
    transactionId?: string;
    userId?: number;
    orderId?: number;
}