# fe/src/app/api/user/model/pageableObject.ts
/**
 * Shop App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SortObject } from './sortObject';


export interface PageableObject { 
    offset?: number;
    sort?: Array<SortObject>;
    unpaged?: boolean;
    pageNumber?: number;
    paged?: boolean;
    pageSize?: number;
}