# fe/src/app/api/user/model/paymentResponse.ts
/**
 * Shop App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentResponse { 
    id?: number;
    orderId?: number;
    userId?: number;
    guestId?: string;
    status?: PaymentResponseStatusEnum;
    amount?: number;
    method?: PaymentResponseMethodEnum;
    transactionId?: string;
    paidAt?: string;
    paymentUrl?: string;
}
export enum PaymentResponseStatusEnum {
    SUCCESS = 'SUCCESS',
    PENDING = 'PENDING',
    PAID = 'PAID',
    FAILED = 'FAILED',
    REFUNDED = 'REFUNDED',
    CANCELLED = 'CANCELLED'
};
export enum PaymentResponseMethodEnum {
    COD = 'COD',
    VNPAY = 'VNPAY'
};